+++ paimon-filesystems/paimon-s3-impl/src/main/java/org/apache/paimon/s3/S3MultiPartUpload.java
@@ -78,7 +78,7 @@ public class S3MultiPartUpload
 
     @Override
     public PartETag uploadPart(
-            String objectName, String uploadId, int partNumber, File file, long byteLength)
+            String objectName, String uploadId, int partNumber, File file, int byteLength)
             throws IOException {
         final UploadPartRequest uploadRequest =
                 s3accessHelper.newUploadPartRequest(
